refactor:  added enum and deleted logs
