+++ mobile/src/features/auth/by-email/model/types.ts
@@ -1,5 +1,7 @@
+import { Role } from '@/shared/libs/types';
+
 export interface AuthByEmailProps {
   email: string;
   password: string;
-  role: string;
+  role: Role;
 }
