+++ mobile/src/features/auth/by-phone/model/types.ts
@@ -1,5 +1,7 @@
+import { Role } from '@/shared/libs/types';
+
 export interface AuthByPhoneProps {
   phone_number: string;
   password: string;
-  role: string;
+  role: Role;
 }
