+++ mobile/src/features/auth/registration/model/types.ts
@@ -1,8 +1,10 @@
+import { Role } from '@/shared/libs/types';
+
 export interface RegistrationProps {
   first_name: string;
   last_name: string;
   email: string;
   password: string;
   phone_number: string;
-  role: string;
+  role: Role;
 }
