+++ mobile/src/features/auth/registration/ui/registrationForm.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Text, View } from 'react-native';
+import { StackActions } from '@react-navigation/native';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { Button } from '@ui-kitten/components';
@@ -11,12 +12,11 @@ import { useAppNavigation } from '@/shared/libs/useAppNavigation';
 import { Input } from '@/shared/ui/input';
 import { Select } from '@/shared/ui/select';
 import { ErrorText } from '@/shared/ui/errorText';
+import { Role } from '@/shared/libs/types';
 
 import { schema } from '../model/validation';
 import { RegistrationProps } from '../model/types';
-import { registration } from '../api';
-
-import Reactotron from 'reactotron-react-native';
+import { registration } from '../api/api';
 
 export const RegistrationForm = () => {
   const {
@@ -35,18 +35,16 @@ export const RegistrationForm = () => {
     setUser: state.setUser,
   }));
 
-  const Roles = ['Customer', 'Vendor'];
+  const Roles = [Role.CUSTOMER, Role.VENDOR];
 
   const onPressSend: SubmitHandler<RegistrationProps> = (formData) => {
     registration(formData)
       .then((data) => {
         setUser(data.user, data.access_token, formData.role);
-        navigation.navigate('Account', { screen: 'AccountScreen' });
-        Reactotron.log(data);
+        navigation.dispatch(StackActions.replace('AccountScreen'));
       })
       .catch((error) => {
         setError('root', { type: 'server', message: 'Такой пользователь уже существует!' });
-        Reactotron.log(error);
       });
   };
